requirements.txt
langchain
langgraph
langchain-openai
langchain-community
chromadb
tiktoken
python-dotenv
